Issue #66:  Convert new V2 REST API to return List<NotificationEntry> instead of NotificationResponse
